Multiple attachments! + Moved session module into it's own file. Updated README.md
