+++ claude2_api/client.py
@@ -2,137 +2,15 @@ import os
 import re
 import json
 import uuid
-import shutil
-import platform
 import requests
-import screeninfo
 import mimetypes
-import selenium
 from time import time
 from datetime import datetime
 from tzlocal import get_localzone
-from selenium.webdriver.firefox.options import Options as FirefoxOptions
 from dataclasses import dataclass
-
-# Selenium module
-
-
-def __cleanup_resources():
-    if platform.system() == "Windows":
-        folder_path = os.path.join(os.environ["LOCALAPPDATA"], "Temp")
-    elif platform.system() == "Linux":
-        folder_path = "/tmp"
-    else:
-        return
-
-    for filename in os.listdir(folder_path):
-        full_path = os.path.join(folder_path, filename)
-        if os.path.isdir(full_path):
-            for dirPrefix in ["tmp", "rust_mozpr"]:
-                if filename.startswith(dirPrefix):
-                    shutil.rmtree(full_path)
-                    break
-
-
-def __get_firefox_options(
-    firefox_profile: str = "", headless: bool = False, private_mode: bool = False
-) -> selenium.webdriver.firefox.options.Options:
-    """Returns chrome options instance with given configuration set"""
-    options = FirefoxOptions()
-    options.profile = (
-        __get_default_firefox_profile() if not firefox_profile else firefox_profile
-    )
-
-    if headless:
-        monitor = screeninfo.get_monitors()[0]
-        options.add_argument("--headless")
-        options.add_argument(f"--window-size={monitor.width},{monitor.height}")
-        options.add_argument("--start-maximized")
-        options.set_preference("media.volume_scale", "0.0")
-
-        # Opt
-        options.set_preference("browser.cache.disk.enable", False)
-        options.set_preference("browser.cache.memory.enable", False)
-        options.set_preference("browser.cache.offline.enable", False)
-        options.set_preference("network.http.use-cache", False)
-
-    if private_mode:
-        options.set_preference(
-            "browser.privatebrowsing.autostart", True
-        )  # Start in private mode
-
-    return options
-
-
-def __get_firefox_webdriver(
-    *args, use_selwire: bool = False, **kwargs
-) -> selenium.webdriver:
-    """Constructor wrapper for Firefox webdriver"""
-
-    if platform.system() == "Windows":
-        # Check if firefox is in PATH
-        DEFAULT_WINDWOS_FIREFOX_PATH = "C:\\Program Files\\Mozilla Firefox"
-        if (
-            not shutil.which("firefox")
-            and os.environ["PATH"].find(DEFAULT_WINDWOS_FIREFOX_PATH) == -1
-        ):
-            os.environ["PATH"] += f";{DEFAULT_WINDWOS_FIREFOX_PATH}"
-
-    if use_selwire:
-        from seleniumwire import webdriver
-
-        return webdriver.Firefox(*args, **kwargs)
-    return selenium.webdriver.Firefox(*args, **kwargs)
-
-
-def __linux_default_firefox_profile_path() -> str:
-    profile_path = os.path.expanduser("~/.mozilla/firefox")
-
-    if not os.path.exists(profile_path):
-        raise RuntimeError(f"\nUnable to retrieve {profile_path} directory")
-
-    for entry in os.listdir(profile_path):
-        if entry.endswith(".default-release"):
-            return os.path.join(profile_path, entry)
-    return None
-
-
-def __win_default_firefox_profile_path() -> str:
-    profile_path = os.path.join(os.getenv("APPDATA"), "Mozilla\Firefox\Profiles")
-    for entry in os.listdir(profile_path):
-        if entry.endswith(".default-release"):
-            return os.path.join(profile_path, entry)
-    return None
-
-
-def __get_default_firefox_profile() -> str:
-    if platform.system() == "Windows":
-        return __win_default_firefox_profile_path()
-    elif platform.system() == "Linux":
-        return __linux_default_firefox_profile_path()
-    return ""
-
+from .session import SessionData
 
 # Claude client module
-
-
-@dataclass(frozen=True)
-class SessionData:
-    """
-    This session class will be passed to `ClaudeAPIClient` constructor.
-
-    It can be auto generated by having a working login in Firefox, and geckodriver installed, calling `get_session_data()`
-    with the Firefox profile path, or the default one if omitted.
-    """
-
-    cookie: str
-    """
-    The entire Cookie header string value
-    """
-    user_agent: str
-    """
-    Browser User agent
-    """
 
 
 @dataclass(frozen=True)
@@ -155,45 +33,7 @@ class SendMessageResponse:
     """
 
 
-def get_session_data(profile: str = "", quiet: bool = False) -> SessionData | None:
-    """
-    Retrieves Claude session data
-
-    This function requires a profile with Claude login and geckodriver installed!
-
-    The default Firefox profile will be used, if the profile argument was not overwrited.
-    """
-
-    __BASE_CHAT_URL = "https://claude.ai/chats"
-    if not profile:
-        profile = __get_default_firefox_profile()
-
-    if not quiet:
-        print(f"\nRetrieving {__BASE_CHAT_URL} session cookie from {profile}")
-
-    __cleanup_resources()
-    opts = __get_firefox_options(firefox_profile=profile, headless=True)
-    driver = __get_firefox_webdriver(options=opts)
-    try:
-        driver.get(__BASE_CHAT_URL)
-
-        driver.implicitly_wait(3)
-        user_agent = driver.execute_script("return navigator.userAgent")
-        if not user_agent:
-            raise RuntimeError("\nCannot retrieve UserAgent...")
-
-        cookies = driver.get_cookies()
-
-        cookie_string = "; ".join(
-            [f"{cookie['name']}={cookie['value']}" for cookie in cookies]
-        )
-        return SessionData(cookie_string, user_agent)
-    finally:
-        driver.quit()
-        __cleanup_resources()
-
-
-class MessageRateLimitHit(Exception):
+class MessageRateLimitError(Exception):
     def __init__(self, resetTimestamp: int, *args: object) -> None:
         super().__init__(*args)
         self.resetTimestamp: int = resetTimestamp
@@ -236,7 +76,7 @@ class ClaudeAPIClient:
                 if name == "sessionKey":
                     return f"{name}={value}"
         raise RuntimeError(
-            "\nCannot retrieve session cookie!\nCheck Claude login in Firefox profile..."
+            "Cannot retrieve session cookie!\nCheck Claude login in Firefox profile..."
         )
 
     def __get_organization_id(self) -> str:
@@ -281,7 +121,7 @@ class ClaudeAPIClient:
         mime_type, _ = mimetypes.guess_type(f"file.{extension}")
         return mime_type or "application/octet-stream"
 
-    def __prepare_file_attachment(self, fpath: str) -> dict:
+    def __prepare_file_attachment(self, fpath: str) -> dict | None:
         content_type = self.__get_content_type(fpath)
         if content_type == "text/plain":
             return self.__prepare_text_file_attachment(fpath)
@@ -313,7 +153,30 @@ class ClaudeAPIClient:
             response = requests.post(url, headers=headers, files=files)
             if response.status_code == 200:
                 return response.json()
-            return {}
+        print(
+            f"\n[{response.status_code}] Unable to prepare file attachment -> {fpath}"
+        )
+        return None
+
+    def __check_file_attachments_paths(self, path_list: list[str]):
+        __FILESIZE_LIMIT = 20971520  # 20 MB
+        if not path_list:
+            return
+
+        if len(path_list) > 5:  # No more than 5 attachments
+            raise ValueError("Cannot attach more than 5 files!")
+
+        for path in path_list:
+            # Check if file exists
+            if not os.path.exists(path) or not os.path.isfile(path):
+                raise ValueError(f"Attachment file does not exists -> {path}")
+
+            # Check file size
+            _size = os.path.getsize(path)
+            if _size > __FILESIZE_LIMIT:
+                raise ValueError(
+                    f"Attachment file exceed file size limit by {_size-__FILESIZE_LIMIT} out of 20MB -> {path}"
+                )
 
     def create_chat(self) -> str | None:
         """
@@ -372,6 +235,9 @@ class ClaudeAPIClient:
         return response.status_code == 204
 
     def get_all_chat_ids(self) -> list[str]:
+        """
+        Retrieve a list with all created chat_ids, empty list if no chat is found.
+        """
         url = f"{self.__BASE_URL}/api/organizations/{self.organization_id}/chat_conversations"
 
         headers = {
@@ -424,7 +290,11 @@ class ClaudeAPIClient:
         return all([self.delete_chat(chat_id) for chat_id in chats])
 
     def send_message(
-        self, chat_id: str, prompt: str, attachment_path: str = "", timeout: int = 240
+        self,
+        chat_id: str,
+        prompt: str,
+        attachment_paths: list[str] = None,
+        timeout: int = 240,
     ) -> SendMessageResponse:
         """
         Send message to `chat_id` using specified `prompt`, loading `attachment_path` if present.
@@ -434,20 +304,24 @@ class ClaudeAPIClient:
         - `status_code` integer field,
         - `error_response` dictionary field, which will be filled in case of errors for inspections purposes.
         """
-        if attachment_path and (
-            not os.path.exists(attachment_path) or not os.path.isfile(attachment_path)
-        ):
-            raise ValueError(f"\nInvalid attachment path -> {attachment_path}")
 
-        attachment = []
-        if attachment_path:
-            attachment = [self.__prepare_file_attachment(attachment_path)]
+        self.__check_file_attachments_paths(attachment_paths)
+
+        attachments = []
+        if attachment_paths:
+            attachments = [
+                a
+                for a in [
+                    self.__prepare_file_attachment(path) for path in attachment_paths
+                ]
+                if a
+            ]
 
         url = f"{self.__BASE_URL}/api/append_message"
 
         payload = json.dumps(
             {
-                "attachments": attachment,
+                "attachments": attachments,
                 "completion": {
                     "model": "claude-2",
                     "prompt": prompt,
@@ -496,7 +370,7 @@ class ClaudeAPIClient:
             decoded_data = response.content.decode("utf-8")
             data = json.loads(decoded_data)
             if data and "error" in data and "resets_at" in data["error"]:
-                raise MessageRateLimitHit(int(data["error"]["resets_at"]))
+                raise MessageRateLimitError(int(data["error"]["resets_at"]))
 
         err = {}
         if not answer:
